Draft: move functions to draftgeoutils.faces
